style: dashboard user interface
