+++ frontend/src/components/layout/Headers.tsx
@@ -3,11 +3,11 @@ import React from 'react'
 
 function Headers() {
   return (
-    <div className='flex justify-between items-center bg-[#1a1a1a] px-10 py-7 border-b border-black '>
+    <div className='flex justify-between items-center  px-10 py-7 '>
         <h1>Logo</h1>
-        <button className='bg-[#ededed] text-[#1a1a1a] py-3 px-6 rounded-full cursor-pointer hover:bg-white'>
+        <button className='bg-[#a42921] text-[white] py-3 px-6 rounded-full cursor-pointer hover:bg-[#b02015]'>
         <Link href='/auth/signin'>
-          Signin
+        Sign In
         </Link>
         </button>
     </div>
